Change starting balance, tweak text
So the starting balance *does* matter. Not for how the pool responds,
that the proportionality constant, but it does matter for how long the
pool can remain solvent. Unlike a constant-product AMM, you can actually
completely drain one side of the pool.

So maybe in a future version, I should let the initial amount be
specified, and compute the proportionality constraint from it so that
the pool can remain solvent up to say 99.5%.