Use Atkinson Hyperlegible for everything
I do love Work Sans, especially the quirky ultrabold. That's what gave
the design its aethetic. It works well for the Hack-o-matic, it can be
quirky, it can be bold. But the Predict-o-matic, I want people to use it
in more serious settings, it could look a bit more professional. Though
I do want it to still have character. The monochrome theme does that,
and one thing that with Atkinson also works really well is to use big
text, *not* too bold, and then have lines of roughly the same thickness
as the text. Like the percentage slider.

Where Atkinson Hyperlegible is lacking though, is that the bold does not
go bold enough. It was working really well for the h3s with Work Sans,
the page layout was very clear. With less bold text, I need to use more
whitespace to clarify the layout.

I think this new one does work, I especially like the stylized parts
where the text is a bit thinner, because Atkinson is more geometric.
It reminds me of something, not sure what. Maybe then NS train schedules
in black on bright yellow, or the monochrome station clocks. It feels
like an older aesthetic, but I like it.