Highlight user and system in participant table
To make it easier to find yourself in the table, put a marker in front
of it. To make it clearer that the system user is special, put it in
italics, and make it lowercase like other usernames to avoid distraction.

Alternatively though, I'd like to continue rendering system in all caps,
but add some letter spacing to make it look nicer. I'm on the fence, but
for now italics it is.

For the user itself, I also played with asterisk instead of guillemet,
but that may signal "there is something special here", and it was
demanding a bit more attention, I think the arrow is nice and subtle,
but still helpful.