Play around with style of system line in table
I also wanted to add an encircled question mark with a link to the help
page, but it starts to look too cluttered, I'm not entirely happy with
it. Maybe I'll revisit it later. Or maybe it doesn't go here but I it
does fit in in other places?

Anyway, both letter spacing + italic looks fine too for the system
account.