Design a favicon
This was done in Inkscape, and then I cleaned up the svg manually in an
editor to remove redundant stuff. It could be improved by rounding the
coordinates, but meh.

The percent sign is from Atkinson Hyperlegible. I only had the bold
variant on my machine, I don't have Hyperlegible Next locally, but the
regular bold was not bold enough for an icon, so I made it bolder by
adding a stroke to the text.