Prepare for serving a favicon
To serve the favicon, we need to add a <link> tag to the head. But the
path for the favicon will depend on the server prefix. So now the head
needs to know the server prefix. But now every error page needs to get
this prefix somehow, so we need to pipe it through. The diff is on the
larger side, but it's just a refactor.

Oh this also adds a function to load the favicon and I can't be bothered
to split that out right now.