Add schema versions, record after balance on transfer
I want to record the after balance to make it easier to graph
probabilities over time. However, that's a breaking change to the
schema, so now I also need a way to do migrations, and this one is
immediately somewhat tricky, because you need to replay all
transactions, which is easier imperatively. Or actually, now that I'm
writing this, maybe it's not too bad with sums?