Add a migration to augment transfers with balance
This updates the schema to v2. A sanity check shows that it _mostly_
works, but I have some accounts that show a delta:

    select
      accounts.id,
      accounts.balance - transfers.from_balance_after,
      accounts.*,
      transfers.*
    from
      accounts,
      transfers
    WHERE
      accounts.id = transfers.from_account_id
    order by
      accounts.id asc,
      transfers.id desc;

Mostly the delta is 0 for the first row per account, except I have some
from my production database where this is not the case! What gives? I
need to debug this further.