Tweak index page summary appearances again
Adding the outcome label does make it too cluttered in my opinion, I
like keeping just the percentages even if it is a bit less clear when
the outcome is not "Yes"/"No", because so far we only have one question
where it's not and for the others a percentage makes sense.

Also, "resolved" should go above the label, it just makes more sense to
read that way even if visually it's not as clear.