Overhaul balances table
This is the third iteration on this. The first insight: it's better to
have a separate balances table than to report the post balance in the
transfers table, because the balances table can be uniform, whereas
transfers has two columns (from and to account id) that you need to
check.

The second insight: we want the post balance after an *event*, not after
a transfer. When an event consists of multiple transfers, we do not care
about intermediate balances because an event is one transaction, and the
intermediate balances are not observable.

I also confirmed on a backup of the production ("production") database
that the final balances computed by the migration match the balances
stored in the accounts table.

    select
      *,
      (
        select balance_after
        from balances b
        where b.account_id = a.id
        order by id desc
        limit 1
      ) - a.balance as diff
    from
      accounts a
    where
      diff <> 0;

This query returns no results. Without 'diff <> 0', it does return the
proper account balances for every account. Also eyeballing the balances
table for a single account joined with the mutating events, it looks at
least plausible for my own trades.