Explain payment in the trade UI
Now with the one slider, you can both trade existing shares, and if
that's not enough to cover the intended trade, do an additional deposit.
(This is implemented in the UI, not yet in the backend.) The trade
widget shows the total proceeds of what you buy (from trade + deposit),
plus how you pay for it (from trade and/or deposit). So the buy sums to
sell + deposit, I think that makes more intuitive sense.