Write about Kelly criterion
I am trying to apply the Kelly Criterion and show the Kelly fraction for
a given bet. Then I realized that I don't fully understand it. After
poking at ChatGPT a bit I understand it better now, but the ultimate
test for whether you understand something, is whether you can explain it
to others, so that's what I'm writing up here.

I wanted to start by writing that the gambler's formula does not work,
but now that I wrote it, it seems to work fine. So now my next step is
to figure out if this is equivalent to the log-wealth maximization ...