Require confirmation to spend >1/2 of the balance
We had some people in practice who went all-in and then made significant
losses or got ruined. And that's even with the explicit deposit. Now
that it's only "move slider, click trade", I think there is more risk
of people accidentally spending large amounts of their balance. So add
a small safeguard to prevent it from happening in a single click. Maybe
the message will also make people realize that spending this much in a
single trade is not a normal thing to do, more so than the deposit field
which had no safeguard against that.